Add tool to tally pieRule results
Here, "pie rule" is a modest variant to Chess (or to any abstract
strategy game):

* Player one chooses White’s first move
* Player two chooses whether to play White or Black

In this case, we want to know the win/lose/draw numbers for White
for any legal first move.  So, we play a bunch of 20-game sets, where
we consider all of White’s regular Chess possible first moves (this
would be 28-game sets with most Capablanca variants), and we see the
White win/lose/draw ratio.