Congo: Add original congo.zip and congo-megapawn files
As per the copyright for congo.zip, we can redistribute as long
as the .zip file is not modified.  That in mind, we have the original
congo.zip file, as well as a script and files to modify Congo to give
it more dynamic and less drawish games.

The rule changes are:

* A1/B1 and F1/G1 have islands: One will not drown on these squares.
* The pawns are now “mPawns”: Once after the river, they can *not* go
  backwards, but can more or capture one square forwards or sideways.
* When a pawn promotes, it becomes a “megapawn”: It may move or capture
  any space within two squares.  This way, like Chess, a pawn promotion
  will usually be a win for the player with the promoted pawn.
* The opening setup has been changed to one which is less passive and
  which can be played using an ordinary Chess board and pieces.