Congo for Zillione: Elephants.
We’re seeing a 20-40% draw rate with Zillions at 7-ply or
3 minutes, but, then again, Chess itself has about a 20-30% draw
rate under the same circumstances.

The issue I’m seeing is that Congo really doesn’t encourage interaction
between the players.  With everyone having only short-range pieces, and
with the river making it harder to cross (even with only three deadly
squares after I put safely islands on A4/B4/F4/G4), I’m now understanding
why Chinese Chess has the Rook and the Cannon, so it’s easier to reach
to the other side of the river.

The games are like this: White and Black passively set up the pieces
on their side, and it takes a while for there to be real interaction
between the pieces.  Also, the pieces are too weak to readily capture
the Lion until the late midgame, which is too un-Chesslike for me.  I
prefer Mad Queen Chess over Shantraj-like games.