SMC new rule: First asymmetrical move must be on White’s right
I have added this rule to the Congo readme file which describes
Simple Megapawn Congo (which has a mirror image opening setup).

This rule is based on a convention Xiangqi appears to have, and
allows moves in opening books to always look like moves actually
played.

This rule does not apply to variants with asymmetrical setups.