TurtleShell.zrf: Add ability for rooks to capture pieces
There isn’t any way to do this except use the example code
Lefler/Mallett provided.

As per Google vs. Oracle, this is “transformative”, to quote that
Supreme Court decision:

	Google copied only what was needed to allow programmers to work
	in a different computing environment without discarding a portion
	of a familiar programming language

Here, I am copying only what is needed to work with a different (i.e.
open content) ZRF without discarding a portion of a familiar game
definition interface.

Micro Star v. FormGen Inc. does not apply here; the issue there was
that a data file could created a different “audiovisual display” of
copyrighted material.  Chess, from before 1500 or so, is not copyrighted;
Turtle Shell Chess, invented by me, is not, to the best of my ability,
coprighted.  None of the graphics, which are derived from 19th century
works or created by myself, are, to the best of my ability, copyrighted.
The data file is in, no way, generating a copyrighted “audiovisual
display”; the Zillions program here is used solely to facilitate the
display of non-copyrighted material and to determine which move to make
when playing the variant.