Done calculating all possible 2x2 short range pieces.
There are, with just leapers, 16,452,080 pieces which can reach
all 64 squares; there are only 325,135 colorbound pieces.

If we take those 16,452,080 and optionally allow to make a given
leap move a “ride” move (knightrider, a rook is a wazir-rider, etc.),
we have 282,232,643,280 possible pieces, just over 2 to the power of
38 (274,877,906,944).

This means an 8x8 board with random non-colorbound pieces and using
a standard chess set has, would you believe, some
6,344,961,231,517,063,209,074,884,200,517,463,972,290,560,000 possible
variants (the pawns and kings can keep their moves), just over 2 to
the power of 152.

If we get fairy pieces (USCFsales.com has 10 different tournament board
compatible fairy pieces for sale as I type this), or if we put a checker
underneath one of the rooks/knights/bishops, we have over 2 ** 266
possible chess variants, a number which needs 81 columns to write.