knight connections (Turtle shell): Slightly revise NW/NE/SW/SE
The issue is this: On an infinite board, A3-A4-A5-A6 and H3-H4-H5-H6 would
be a separate file; the only reason why these eight squares are only in
one file in Turtle Shell is because those extra files have, until now,
not changed any moves.

With the NW/NE/SW/SE rules, that has changed.