TurtleShell.zrf: I think this kind of diagonal move works
This new diagonal move is color-bound: Pieces on squares stay on squares,
pieces on triangles stay on triangles.

On triangles, a piece moves across a group of three triangles
together, then it moves across a corner to the end triangle in
another group of tree triangles.

With squares, if on a tilted square tile, it moves to a group of
two straight squares it shares a corner with.  If on a straight square
tile, it moves to the other straight square next to it, then moves
to the tilted square it shares a corner with.

The exact paths can be seen in the diff.